requirements.txt
aws-cdk.core>=1.132.0
aws-cdk.aws_kinesis>=1.132.0
aws-cdk.aws_glue>=1.132.0
aws-cdk.aws_s3>=1.132.0